split lib code
